productis en base de datos con precio base usd y migracion activa

+++ backend/facturacion/models.py
@@ -64,7 +64,7 @@ class DetalleFactura(models.Model):
     total = models.DecimalField(max_digits=15, decimal_places=2)
     porcentaje_ganancia = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     precio_compra_usd = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
-    unidades_paquete = models.IntegerField(default=1)
+    unidades_paquete = models.DecimalField(max_digits=10, decimal_places=2, default=1)
     aplicarIva = models.BooleanField(default=False)
     precio_base_usd = models.DecimalField(max_digits=10, decimal_places=2, null=True, blank=True)
     tipo_tasa = models.CharField(max_length=10, null=True, blank=True)
